refactor Step 06, update dependencies
