+++ 06-ssr-helmet/package.json
@@ -17,18 +17,18 @@
     "> 1%"
   ],
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.2",
-    "babel-preset-env": "^1.6.0",
-    "babel-preset-react": "^6.24.1",
+    "@babel/cli": "^7.1.5",
+    "@babel/core": "^7.1.6",
+    "@babel/node": "^7.0.0",
+    "@babel/preset-env": "^7.1.6",
+    "@babel/preset-react": "^7.0.0",
     "cross-env": "^5.0.5",
-    "eslint": "^4.9.0",
-    "eslint-config-airbnb": "^16.1.0",
+    "eslint": "^5.9.0",
+    "eslint-config-airbnb": "17.1.0",
     "eslint-plugin-compat": "^1.0.4",
-    "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-jsx-a11y": "^6.0.2",
-    "eslint-plugin-react": "^7.4.0",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-jsx-a11y": "^6.1.2",
+    "eslint-plugin-react": "^7.11.1",
     "husky": "^0.14.3",
     "nodemon": "^1.12.1",
     "pm2": "^2.7.1",
@@ -38,7 +38,8 @@
     "webpack-dev-server": "^2.9.1"
   },
   "dependencies": {
-    "babel-polyfill": "^6.26.0",
+    "@babel/polyfill": "^7.0.0",
+    "babel-loader": "^8.0.4",
     "compression": "^1.7.1",
     "express": "^4.16.1",
     "isomorphic-fetch": "^2.2.1",
