+++ 06-ssr-helmet/src/shared/pages/Home/index.jsx
@@ -2,7 +2,7 @@ import 'isomorphic-fetch';
 import React, { Component } from 'react';
 import Helmet from 'react-helmet';
 
-import { firstEndpointRoute } from '../../../shared/routes';
+import { firstEndpointRoute } from '../../routes';
 
 import Button from '../../components/Button';
 
@@ -21,6 +21,7 @@ class Home extends Component {
     this.clickHandler = this.clickHandler.bind(this);
     this.clickHandlerAsync = this.clickHandlerAsync.bind(this);
   }
+
   clickHandler() {
     // eslint-disable-next-line no-console
     console.log('Button clicked');
@@ -33,6 +34,7 @@ class Home extends Component {
       buttonText: `Good Morning again at ${time.getHours()}:${time.getMinutes()}:${time.getSeconds()}`,
     });
   }
+
   clickHandlerAsync() {
     // eslint-disable-next-line no-console
     console.log('Async Button clicked');
@@ -59,13 +61,15 @@ class Home extends Component {
         })
     );
   }
+
   render() {
+    const { buttonText, buttonTextAsync } = this.state;
     return (
       <div>
         <Helmet title="Home" meta={[{ name: 'description', content: 'Home Page description' }]} />
         <h1>Home</h1>
-        <Button text={this.state.buttonText} onButtonClick={this.clickHandler} />
-        <Button text={this.state.buttonTextAsync} onButtonClick={this.clickHandlerAsync} />
+        <Button text={buttonText} onButtonClick={this.clickHandler} />
+        <Button text={buttonTextAsync} onButtonClick={this.clickHandlerAsync} />
       </div>
     );
   }
